refactor: Props 타입을 Flex.types.ts로 분리하여 순환 참조 문제 해결
